Clean up payload parser
This adds a few missing fields and cleans up the implementation
and missing tests related to converting a ReportableError into an
ErrorPayload.

This does not include any validation as of yet since the ReportableError
type is so type safe at this point. When I add support for other types of
data to be uploaded, that can be extended.